Port NestedElementQueryInterface
