+++ src/Element/Queries/Contracts/NestedElementQueryInterface.php
@@ -1,0 +1,219 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CraftCms\Cms\Element\Queries\Contracts;
+
+use craft\base\ElementInterface;
+
+/**
+ * NestedElementQueryInterface defines the common interface to be implemented by element query classes
+ * which can query for nested elements.
+ */
+interface NestedElementQueryInterface extends ElementQueryInterface
+{
+    /**
+     * Narrows the query results based on the field the {elements} are contained by.
+     *
+     * Possible values include:
+     *
+     * | Value | Fetches {elements}…
+     * | - | -
+     * | `'foo'` | in a field with a handle of `foo`.
+     * | `['foo', 'bar']` | in a field with a handle of `foo` or `bar`.
+     * | a [[craft\fields\Matrix]] object | in a field represented by the object.
+     *
+     * ---
+     *
+     * ```twig
+     * {# Fetch {elements} in the Foo field #}
+     * {% set {elements-var} = {twig-method}
+     *   .field('foo')
+     *   .all() %}
+     * ```
+     *
+     * ```php
+     * // Fetch {elements} in the Foo field
+     * ${elements-var} = {php-method}
+     *     ->field('foo')
+     *     ->all();
+     * ```
+     *
+     * @param  mixed  $value  The property value
+     * @return static self reference
+     */
+    public function field(mixed $value): static;
+
+    /**
+     * Narrows the query results based on the field the {elements} are contained by, per the fields’ IDs.
+     *
+     * Possible values include:
+     *
+     * | Value | Fetches {elements}…
+     * | - | -
+     * | `1` | in a field with an ID of 1.
+     * | `'not 1'` | not in a field with an ID of 1.
+     * | `[1, 2]` | in a field with an ID of 1 or 2.
+     * | `['not', 1, 2]` | not in a field with an ID of 1 or 2.
+     *
+     * ---
+     *
+     * ```twig
+     * {# Fetch {elements} in the field with an ID of 1 #}
+     * {% set {elements-var} = {twig-method}
+     *   .fieldId(1)
+     *   .all() %}
+     * ```
+     *
+     * ```php
+     * // Fetch {elements} in the field with an ID of 1
+     * ${elements-var} = {php-method}
+     *     ->fieldId(1)
+     *     ->all();
+     * ```
+     *
+     * @param  mixed  $value  The property value
+     * @return static self reference
+     */
+    public function fieldId(mixed $value): static;
+
+    /**
+     * Narrows the query results based on the primary owner element of the {elements}, per the owners’ IDs.
+     *
+     * Possible values include:
+     *
+     * | Value | Fetches {elements}…
+     * | - | -
+     * | `1` | created for an element with an ID of 1.
+     * | `[1, 2]` | created for an element with an ID of 1 or 2.
+     *
+     * ---
+     *
+     * ```twig
+     * {# Fetch {elements} created for an element with an ID of 1 #}
+     * {% set {elements-var} = {twig-method}
+     *   .primaryOwnerId(1)
+     *   .all() %}
+     * ```
+     *
+     * ```php
+     * // Fetch {elements} created for an element with an ID of 1
+     * ${elements-var} = {php-method}
+     *     ->primaryOwnerId(1)
+     *     ->all();
+     * ```
+     *
+     * @param  mixed  $value  The property value
+     * @return static self reference
+     */
+    public function primaryOwnerId(mixed $value): static;
+
+    /**
+     * Sets the [[primaryOwnerId()]] and [[siteId()]] parameters based on a given element.
+     *
+     * ---
+     *
+     * ```twig
+     * {# Fetch {elements} created for this entry #}
+     * {% set {elements-var} = {twig-method}
+     *   .primaryOwner(myEntry)
+     *   .all() %}
+     * ```
+     *
+     * ```php
+     * // Fetch {elements} created for this entry
+     * ${elements-var} = {php-method}
+     *     ->primaryOwner($myEntry)
+     *     ->all();
+     * ```
+     *
+     * @param  ElementInterface  $primaryOwner  The primary owner element
+     * @return static self reference
+     */
+    public function primaryOwner(ElementInterface $primaryOwner): static;
+
+    /**
+     * Narrows the query results based on the owner element of the {elements}, per the owners’ IDs.
+     *
+     * Possible values include:
+     *
+     * | Value | Fetches {elements}…
+     * | - | -
+     * | `1` | created for an element with an ID of 1.
+     * | `[1, 2]` | created for an element with an ID of 1 or 2.
+     *
+     * ---
+     *
+     * ```twig
+     * {# Fetch {elements} created for an element with an ID of 1 #}
+     * {% set {elements-var} = {twig-method}
+     *   .ownerId(1)
+     *   .all() %}
+     * ```
+     *
+     * ```php
+     * // Fetch {elements} created for an element with an ID of 1
+     * ${elements-var} = {php-method}
+     *     ->ownerId(1)
+     *     ->all();
+     * ```
+     *
+     * @param  mixed  $value  The property value
+     * @return static self reference
+     */
+    public function ownerId(mixed $value): static;
+
+    /**
+     * Sets the [[ownerId()]] and [[siteId()]] parameters based on a given element.
+     *
+     * ---
+     *
+     * ```twig
+     * {# Fetch {elements} created for this entry #}
+     * {% set {elements-var} = {twig-method}
+     *   .owner(myEntry)
+     *   .all() %}
+     * ```
+     *
+     * ```php
+     * // Fetch {elements} created for this entry
+     * ${elements-var} = {php-method}
+     *     ->owner($myEntry)
+     *     ->all();
+     * ```
+     *
+     * @param  ElementInterface  $owner  The owner element
+     * @return static self reference
+     */
+    public function owner(ElementInterface $owner): static;
+
+    /**
+     * Narrows the query results based on whether the {elements}’ owners are drafts.
+     *
+     * Possible values include:
+     *
+     * | Value | Fetches {elements}…
+     * | - | -
+     * | `true` | which can belong to a draft.
+     * | `false` | which cannot belong to a draft.
+     *
+     * @param  bool|null  $value  The property value
+     * @return static self reference
+     */
+    public function allowOwnerDrafts(?bool $value = true): static;
+
+    /**
+     * Narrows the query results based on whether the {elements}’ owners are revisions.
+     *
+     * Possible values include:
+     *
+     * | Value | Fetches {elements}…
+     * | - | -
+     * | `true` | which can belong to a revision.
+     * | `false` | which cannot belong to a revision.
+     *
+     * @param  bool|null  $value  The property value
+     * @return static self reference
+     */
+    public function allowOwnerRevisions(?bool $value = true): static;
+}
