+++ yii2-adapter/legacy/controllers/AppController.php
@@ -10,7 +10,6 @@ namespace craft\controllers;
 use Carbon\CarbonInterval;
 use Craft;
 use craft\base\ElementInterface;
-use craft\elements\db\NestedElementQueryInterface;
 use craft\helpers\App;
 use craft\helpers\Component;
 use craft\helpers\DateTimeHelper;
@@ -25,6 +24,7 @@ use CraftCms\Cms\Cp\Alerts;
 use CraftCms\Cms\Cp\Html\ElementHtml;
 use CraftCms\Cms\Cp\Html\MenuHtml;
 use CraftCms\Cms\Element\Drafts;
+use CraftCms\Cms\Element\Queries\Contracts\NestedElementQueryInterface;
 use CraftCms\Cms\License\License;
 use CraftCms\Cms\Plugin\Plugins;
 use CraftCms\Cms\Shared\Enums\LicenseKeyStatus;
