+++ yii2-adapter/legacy/controllers/ElementsController.php
@@ -10,7 +10,6 @@ namespace craft\controllers;
 use Craft;
 use craft\base\ElementInterface;
 use craft\base\NestedElementInterface;
-use craft\elements\db\NestedElementQueryInterface;
 use craft\events\DefineElementEditorHtmlEvent;
 use craft\services\Drafts;
 use craft\web\Controller;
@@ -33,6 +32,7 @@ use CraftCms\Cms\Element\Exceptions\InvalidElementException;
 use CraftCms\Cms\Element\Exceptions\InvalidTypeException;
 use CraftCms\Cms\Element\Exceptions\UnsupportedSiteException;
 use CraftCms\Cms\Element\Queries\Contracts\ElementQueryInterface;
+use CraftCms\Cms\Element\Queries\Contracts\NestedElementQueryInterface;
 use CraftCms\Cms\Element\Revisions;
 use CraftCms\Cms\FieldLayout\FieldLayoutForm;
 use CraftCms\Cms\FieldLayout\LayoutElements\BaseField;
